+++ datafusion/core/src/physical_optimizer/tests/replace_with_order_preserving_variants.rs
